requirements.txt
boto3==1.35.97
aws-cdk-lib==2.175.1
constructs>=10.0.0,<11.0.0
awscli==1.36.39